+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.36.4
 
+* Fix: relative path detection for workspace members defined in dependencies #773
 * Maintenance: Bump cargo-watch to 0.8.3 #771
 * Maintenance: Bump rust minimal version to 0.65 due to dependencies updates
 * Maintenance: Upgrade dependencies